requirements.txt
requests
python-dotenv
mcp[cli]
fastmcp
httpx